Initial draft of redirectors
This is the first draft of the redirector provisioning and configuration automation. It supports the provisioning of N redirectors, each with their own IP address on the default interface. It installs socat on each of the redirectors, and configures them to redirect all TCP ports 80 and 443 traffic to the specified C2 server on those same ports.